interactive_face_detection, object_detection_demo_ssd_async, smart_cl…

+++ demos/interactive_face_detection_demo/detectors.hpp
@@ -74,8 +74,8 @@ struct FaceDetection : BaseDetection {
     int enquedFrames;
     float width;
     float height;
-    float network_input_width;
-    float network_input_height;
+    size_t network_input_width;
+    size_t network_input_height;
     float bb_enlarge_coefficient;
     float bb_dx_coefficient;
     float bb_dy_coefficient;
